WEB-4025: moved tasks files into dashboard components folder, load campaign server side for dashboard
