+++ app/(candidate)/dashboard/components/DashboardPage.js
@@ -1,6 +1,6 @@
 'use client'
 import DashboardLayout from '../shared/DashboardLayout'
-import { weekRangeFromDate, weeksTill } from 'helpers/dateHelper'
+import { weeksTill } from 'helpers/dateHelper'
 import { useCallback, useEffect, useState } from 'react'
 import { calculateContactGoals } from './voterGoalsHelpers'
 import ElectionOver from './ElectionOver'
@@ -8,22 +8,25 @@ import EmptyState from './EmptyState'
 import { updateUser } from 'helpers/userHelper'
 import { useUser } from '@shared/hooks/useUser'
 import PrimaryResultModal from './PrimaryResultModal'
-import { fetchUserClientCampaign } from 'helpers/fetchUserClientCampaign'
 import LoadingAnimation from '@shared/utils/LoadingAnimation'
 import { VoterContactsProvider } from '@shared/hooks/VoterContactsProvider'
 import { CampaignUpdateHistoryProvider } from '@shared/hooks/CampaignUpdateHistoryProvider'
-import TasksList from '../tasks/components/TasksList'
+import TasksList from './tasks/TasksList'
 
-export default function DashboardPage({ pathname, tasks }) {
+export default function DashboardPage({
+  pathname,
+  tasks,
+  campaign: campaignProp,
+}) {
   const [_, setUser] = useUser()
-  const [campaign, setCampaign] = useState(null)
+  const [campaign, setCampaign] = useState(campaignProp)
   const { pathToVictory: p2vObject, goals, details } = campaign || {}
   const pathToVictory = p2vObject?.data || {}
   const { primaryElectionDate } = details || {}
   const [primaryResultState, setPrimaryResultState] = useState({
     modalOpen: false,
     modalDismissed: false,
-    primaryResult: undefined,
+    primaryResult: campaignProp.details?.primaryResult,
   })
 
   const officeName =
@@ -31,34 +34,18 @@ export default function DashboardPage({ pathname, tasks }) {
       ? details?.otherOffice
       : details?.office
 
-  // TODO: we're only having to do this, because we're caching the user object in the cookie and
-  //  accessing it from there, instead of the source of truth, the DB.
-  //  What we should be doing is fetching the user object from the server on each route change,
-  //  and then we won't have to do this.
-  const updateUserCookie = async () => {
-    const updated = await updateUser()
-    if (updated) {
-      setUser(updated)
-    }
-  }
-
   useEffect(() => {
-    if (campaign) return
+    updateUserCookie()
 
-    loadCampaign()
-
-    async function loadCampaign() {
-      const campaign = await fetchUserClientCampaign()
-      setCampaign(campaign)
-      const storedPrimaryResult = campaign.details?.primaryResult
-
-      setPrimaryResultState({
-        modalOpen: false,
-        modalDismissed: false,
-        primaryResult: storedPrimaryResult,
-      })
-
-      updateUserCookie()
+    // TODO: we're only having to do this, because we're caching the user object in the cookie and
+    //  accessing it from there, instead of the source of truth, the DB.
+    //  What we should be doing is fetching the user object from the server on each route change,
+    //  and then we won't have to do this.
+    async function updateUserCookie() {
+      const updated = await updateUser()
+      if (updated) {
+        setUser(updated)
+      }
     }
   }, [])
 
@@ -84,8 +71,6 @@ export default function DashboardPage({ pathname, tasks }) {
   }
 
   const weeksUntil = weeksTill(resolvedDate)
-
-  const dateRange = weekRangeFromDate(resolvedDate, weeksUntil.weeks)
   const contactGoals = calculateContactGoals(resolvedContactGoal)
 
   const primaryResultCloseCallback = useCallback((selectedResult) => {
@@ -115,19 +100,10 @@ export default function DashboardPage({ pathname, tasks }) {
     }
   }, [])
 
-  const childProps = {
-    campaign,
-    pathname,
-    contactGoals,
-    weeksUntil,
-    dateRange,
-    pathToVictory,
-  }
-
   return (
     <VoterContactsProvider>
       <CampaignUpdateHistoryProvider>
-        <DashboardLayout {...childProps}>
+        <DashboardLayout pathname={pathname} campaign={campaign}>
           {campaign ? (
             <>
               <div>
