+++ app/(candidate)/dashboard/components/tasks/CopyScriptButton.js
