+++ app/(candidate)/dashboard/components/tasks/flows/AddScriptStep/AddScriptStep.js
