+++ app/(candidate)/dashboard/components/tasks/flows/AddScriptStep/AiTemplateSelect.js
