+++ app/(candidate)/dashboard/components/tasks/flows/AddScriptStep/CreateSmSScriptScreen.js
