+++ app/(candidate)/dashboard/components/tasks/flows/AddScriptStep/GenerateLoadingScreen.js
