+++ app/(candidate)/dashboard/components/tasks/flows/CloseConfirmModal.js
