+++ app/(candidate)/dashboard/components/tasks/flows/ScheduleStep.js
@@ -7,7 +7,10 @@ import { useState, useMemo } from 'react'
 import { getDefaultVoterFileName } from 'app/(candidate)/dashboard/voter-records/components/VoterFileTypes'
 import { useSnackbar } from 'helpers/useSnackbar'
 import Button from '@shared/buttons/Button'
-import { LEGACY_TASK_TYPES, TASK_TYPES } from '../../constants/tasks.const'
+import {
+  LEGACY_TASK_TYPES,
+  TASK_TYPES,
+} from '../../../shared/constants/tasks.const'
 
 export default function ScheduleStep({
   onChangeCallback,
