+++ app/(candidate)/dashboard/page.js
@@ -1,6 +1,7 @@
 import pageMetaData from 'helpers/metadataHelper'
 import DashboardPage from './components/DashboardPage'
 import candidateAccess from './shared/candidateAccess'
+import { fetchUserCampaign } from '../onboarding/shared/getCampaign'
 import { apiRoutes } from 'gpApi/routes'
 import { serverFetch } from 'gpApi/serverFetch'
 
@@ -23,7 +24,10 @@ export const metadata = meta
 export default async function Page() {
   await candidateAccess()
 
+  const campaign = await fetchUserCampaign()
   const tasks = await fetchTasks()
 
-  return <DashboardPage pathname="/dashboard" tasks={tasks} />
+  return (
+    <DashboardPage pathname="/dashboard" campaign={campaign} tasks={tasks} />
+  )
 }
