+++ app/(candidate)/dashboard/voter-records/[type]/components/ScheduleCard.js
@@ -2,11 +2,11 @@ import Body2 from '@shared/typography/Body2'
 import H3 from '@shared/typography/H3'
 import Overline from '@shared/typography/Overline'
 import Paper from '@shared/utils/Paper'
-import TaskFlow from 'app/(candidate)/dashboard/tasks/components/flows/TaskFlow'
+import TaskFlow from 'app/(candidate)/dashboard/components/tasks/flows/TaskFlow'
 import {
   LEGACY_TASK_TYPES,
   TASK_TYPES,
-} from 'app/(candidate)/dashboard/tasks/constants/tasks.const'
+} from 'app/(candidate)/dashboard/shared/constants/tasks.const'
 
 export default function ScheduleCard(props) {
   const { type } = props
